% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\docType{data}
\name{counties}
\alias{counties}
\title{Census ata on US counties}
\format{
A data frame with 3143 rows and 8 variables:
\describe{
  \item{id}{GEO.id from original data}
  \item{state}{state in which the county is located}
  \item{county}{name of the county}
  \item{population}{population of the county}
  \item{housingUnits}{housing units in the county}
  \item{totalArea}{Area in square miles - Total area}
  \item{waterArea}{Area in square miles - Water area}
  \item{landArea}{Area in square miles - Land area}
}
}
\source{
http://factfinder.census.gov/bkmk/table/1.0/en/DEC/10_SF1/GCTPH1.US05PR (link now dead), system.file("data-raw", "makeCounties.R", package = "vipor")
}
\usage{
counties
}
\description{
A dataset containing data from the US census burea
}
\references{
\url{https://web.archive.org/web/20150326040847/https://www.census.gov/prod/cen2010/cph-2-1.pdf}
}
\keyword{datasets}
