% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{generatePermuteString}
\alias{generatePermuteString}
\title{Generate a permutation string meeting Tukey criteria}
\usage{
generatePermuteString(nReps = 20, n = 5)
}
\arguments{
\item{nReps}{number of permutations to concatenate}

\item{n}{permutations from 1 to n}
}
\value{
a vector of nReps*n integers giving concatenated permutations
}
\description{
Find a random string of concatenated permutations of 1:n fulfilling Tukey's criteria that there are no runs of 3 or more increases or decreases in a row. Tukey just uses the default n=5.
}
\examples{
tukeyPermutes()
tukeyPermutes(6,3)
}

