% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/violinPoint.R
\name{offsetSingleGroup}
\alias{offsetSingleGroup}
\title{Offset data to avoid overplotting for a single subgroup of data}
\usage{
offsetSingleGroup(y_subgroup, maxLength = NULL, method = c("quasirandom",
  "pseudorandom", "smiley", "frowney"), nbins = 1000, adjust = 0.5)
}
\arguments{
\item{y_subgroup}{y values for a single group for which offsets should be calculated}

\item{maxLength}{multiply the offset by sqrt(length(y_subgroup)/maxLength) if not NULL. The sqrt is to match boxplot (allows comparison of order of magnitude different ns, scale with standard error)}

\item{method}{method used to distribute the points}

\item{nbins}{the number of points used to calculate density}

\item{adjust}{bandwidth used to adjust the density}
}
\value{
a vector with of x-offsets between -1 and 1 of the same length as y
}
\description{
Arranges data points using a van der Corput sequence, pseudorandom noise or
alternatively positioning extreme values within a band to the left and right to
form "beeswarm" style plots. Returns a vector of the offsets to be used in
plotting.
}

