% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/violinPoint.R
\name{offsetX}
\alias{offsetX}
\title{Offset data to avoid overplotting.}
\usage{
offsetX(y, x, width = 0.4, varwidth = FALSE, adjust = 0.5, nbins = 1000)
}
\arguments{
\item{y}{vector of data points}

\item{x}{a grouping factor for y (optional)}

\item{width}{the maximum spacing away from center for each group of points. Since points are spaced to left and right, the maximum width of the cluster will be approximately width*2 (0 = no offset, default = 0.4)}

\item{varwidth}{adjust the width of each group based on the number of points in the group}

\item{adjust}{bandwidth used to adjust the density}

\item{nbins}{the number of points used to calculate density}
}
\value{
a vector with of x-offsets of the same length as y
}
\description{
Arranges data points using a van der Corput sequence to form "beeswarm" style
plots. Returns a vector of the offsets to be used in plotting.
}
\examples{
## Generate fake data
dat <- list(rnorm(50), rnorm(500), c(rnorm(100), rnorm(100,5)), rcauchy(100))
names(dat) <- c("Normal", "Dense Normal", "Bimodal", "Extremes")

## Plot each distribution with a variety of parameters
par(mfrow=c(4,1), mar=c(2,4, 0.5, 0.5))
sapply(names(dat),function(label) {
  y<-dat[[label]]
  ids <- rep(1:4, each=length(y))

  offsets <- c(
    offsetX(y),  # Default
    offsetX(y, adjust=2),    # More smoothing
    offsetX(y, adjust=0.1),  # Tighter fit
    offsetX(y, width=0.1))   # Less wide

  plot(offsets + ids, rep(y, 4), ylab=label, xlab='', xaxt='n', pch=21, las=1)
  axis(1, 1:4, c("Default", "Adjust=2", "Adjust=0.1", "Width=10\%"))
})
}

