% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{permute}
\alias{permute}
\title{Return all permutations of a vector}
\usage{
permute(vals)
}
\arguments{
\item{vals}{a vector of elements to be permuted}
}
\value{
A list of vectors containing all permutation of the values
}
\description{
Recursively generates all permutations of a vector. The result will be \code{factorial(length(vals))} long so be careful with any longer vectors (e.g. longer than 10).
}
\examples{
permute(letters[1:3])
permute(1:5)
}
\seealso{
\code{\link{sample}}
}

