% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukeyPermutes}
\alias{tukeyPermutes}
\title{Find permutations meeting Tukey criteria}
\usage{
tukeyPermutes(n = 5, limit = 2)
}
\arguments{
\item{n}{permutations from 1 to n}

\item{limit}{the maximum number of increases or decreases in a row}
}
\value{
a list of vectors containing valid permutations
}
\description{
Find all permutations of 1:n fulfilling Tukey's criteria that there are no runs of 3 or more increases or decreases in a row. Tukey just uses the default n=5 and limit=2.
}
\examples{
tukeyPermutes()
tukeyPermutes(6,3)
}

