% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukeyT}
\alias{tukeyT}
\title{Combine multiple permutation strings into one}
\usage{
tukeyT(nReps = 10, base = 5)
}
\arguments{
\item{nReps}{number of permutations to paste together}

\item{base}{generate permutations of integers 1:base}
}
\value{
A nReps*base length vector giving offset positions based on Tukey's algorithm
}
\description{
Combine base+1 permutation strings to generate offsets
}
\examples{
tukeyT()
tukeyT()
tukeyT(5,4)
}

