% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukeyTexture}
\alias{tukeyTexture}
\title{Generate random positions based on Tukey texture algorithm}
\usage{
tukeyTexture(x, jitter = TRUE)
}
\arguments{
\item{x}{the points to be jittered. really only used to calculate length}

\item{jitter}{if TRUE add random jitter to each point}
}
\value{
a vector of length length(x) giving displacements for each corresponding point in x
}
\description{
Generate partly random, partly constrained lateral displacements based on Tukey texture algorithm from Tukey and Tukey 1990
}

